move fallback to utils
