+++ _src/blocks/sticky-navigation/sticky-navigation.js
@@ -1,16 +1,6 @@
 import { STICKY_NAVIGATION_DATASET_KEY } from '../../scripts/lib-franklin.js';
 
-const DSN_FALLBACK = 'https://esm.sh/@repobit/dex-system-design@0.23.30/';
-
-const getDsnBase = () => {
-  try {
-    const map = document.querySelector('script[type="importmap"]');
-    const imports = JSON.parse(map?.textContent || '{}').imports || {};
-    return imports['@repobit/dex-system-design/'] || DSN_FALLBACK;
-  } catch {
-    return DSN_FALLBACK;
-  }
-};
+import { getDsnBase } from '../../scripts/utils/utils.js';
 
 // ─── helpers ───────────────────────────────────────────────────────────────────
